alx llp

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -9,7 +9,7 @@ int main(void)
 {
 int i = 0;
 for (i = 0; i < 10; i++)
-putchar(i);
+putchar((int)i);
 putchar('\n');
 return (0);
 }
